move ElevationProfileChip to correct folder
