+++ test/modules/elevationProfile/components/assistChip/ElevationProfileChip.test.js
@@ -1,9 +1,9 @@
-import { $injector } from '../../../../src/injection';
-import { ElevationProfileChip } from '../../../../src/modules/chips/components/assistChips/ElevationProfileChip';
-import { updateCoordinates } from '../../../../src/store/elevationProfile/elevationProfile.action';
-import { elevationProfileReducer } from '../../../../src/store/elevationProfile/elevationProfile.reducer';
-import { TestUtils } from '../../../test-utils';
-import profileSvg from '../../../../src/modules/chips/components/assistChips/assets/profile.svg';
+import { $injector } from '../../../../../src/injection';
+import { updateCoordinates } from '../../../../../src/store/elevationProfile/elevationProfile.action';
+import { elevationProfileReducer } from '../../../../../src/store/elevationProfile/elevationProfile.reducer';
+import { TestUtils } from '../../../../test-utils';
+import profileSvg from '../../../../../src/modules/elevationProfile/components/assistChip/assets/profile.svg';
+import { ElevationProfileChip } from '../../../../../src/modules/elevationProfile/components/assistChip/ElevationProfileChip';
 
 window.customElements.define(ElevationProfileChip.tag, ElevationProfileChip);
